fix overflow in table and refactore code

+++ include/HashStrInt.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stddef.h>
 #include <stdint.h>
 typedef struct {
   int occupied;
@@ -12,7 +13,7 @@ typedef struct {
   uint32_t size;
 } HashStrInt;
 
-int col_index_reserve(HashStrInt *h, int n_strs);
+int col_index_reserve(HashStrInt *h, size_t n_strs);
 void col_index_destroy(HashStrInt *h);
 int col_index_put(HashStrInt *h, const char *key, int value);
 int col_index_get(const HashStrInt *h, const char *key);
